pyproject.toml
[project]
name = "langchain"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "chromadb>=1.5.8",
    "dashscope>=1.25.17",
    "langchain>=1.2.15",
    "langchain-community>=0.4.1",
    "langchain-huggingface>=1.2.2",
    "langchain-openai>=1.2.1",
    "openai>=2.32.0",
    "python-dotenv>=1.2.2",
    "scipy-stubs~=1.17.1",
    "sentence-transformers>=5.4.1",
]
